pyproject.toml
[project]
name = "elevenlabs-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "elevenlabs[pyaudio]>=2.11.0",
    "python-dotenv>=1.1.1",
    "sounddevice>=0.5.2",
    "fastapi[standard]>=0.111.0",
    "uvicorn[standard]>=0.30.0",
]
